OOP for linked ist and implement dict with last in js - OOP for linked list -> seperate and resuse properties for singly and double linked list - implement dict with last in JS using map and double linked list
